improve test cases to check if the board/lib installation succeeds

+++ test/boardmanager.test.ts
@@ -48,7 +48,7 @@ suite("Arduino: Board Manager.", () => {
         assert.equal(platforms[0].name, "Arduino AVR Boards", "Board Manager should display built Arduino AVR Boards");
         assert.equal(platforms[0].installedVersion, "1.6.18", "Arduino IDE built-in AVR board package version should be 1.6.18");
         assert.equal(platforms[0].rootBoardPath, Path.join(Resources.mockedIDEPackagePath, "arduino", "avr"),
-        "Should be able to index root board path for installed boards");
+            "Should be able to index root board path for installed boards");
     });
 
     test("should be able to load installed platforms", () => {
@@ -65,7 +65,7 @@ suite("Arduino: Board Manager.", () => {
         assert.equal(boardManager.installedBoards.size, 46, "Arduino IDE should contains built-in AVR boards");
         assert.equal(boardManager.installedBoards.has("arduino:avr:yun"), true, "should parse installed boards from Arduino IDE built-in packages");
         assert.equal(boardManager.installedBoards.has("esp8266:esp8266:huzzah"), true,
-        "should parse installed boards from custom packages ($sketchbook/hardware directory)");
+            "should parse installed boards from custom packages ($sketchbook/hardware directory)");
     });
 
     test("should parse boards.txt correctly", () => {
@@ -105,7 +105,16 @@ suite("Arduino: Board Manager.", () => {
         this.timeout(4 * 60 * 1000);
         try {
             // Board Manager: install boards packages.
-            ArduinoContext.arduinoApp.installBoard ("esp8266", "esp8266", "2.3.0", true).then(done, done);
+            ArduinoContext.arduinoApp.installBoard("esp8266", "esp8266", "2.3.0", true).then((result) => {
+                const arduinoSettings = ArduinoContext.arduinoApp.settings;
+                const packagePath = Path.join(arduinoSettings.packagePath, "packages", "esp8266");
+                // check if the installation succeeds or not
+                if (util.directoryExistsSync(packagePath)) {
+                    done();
+                } else {
+                    done(new Error("esp8266 board package install failure, can't find package path :" + packagePath));
+                }
+            });
 
         } catch (error) {
             done(new Error(error));
@@ -117,13 +126,12 @@ suite("Arduino: Board Manager.", () => {
     test("should be able to remove boards packages", () => {
         try {
             // Board Manager: remove boards packages.
-            let packagePath = "/Users/$USER/Library/Arduino15/packages/esp8266";
-            if (os.platform() === "linux") {
-                packagePath = "//home//$USER//.Arduino15//packages//esp8266";
+            const arduinoSettings = ArduinoContext.arduinoApp.settings;
+            const packagePath = Path.join(arduinoSettings.packagePath, "packages", "esp8266");
+            if (util.directoryExistsSync(packagePath)) {
+                ArduinoContext.arduinoApp.uninstallBoard("esp8266", packagePath);
+
             }
-
-            ArduinoContext.arduinoApp.uninstallBoard("esp8266", packagePath);
-
         } catch (error) {
             assert.fail(true, false, new Error(error).message, new Error(error).name);
         }
